tools/golden/pyproject.toml
[project]
name = "omni-golden"
version = "0.1.0"
description = "Golden master testing for omni CLI"
requires-python = ">=3.10"
dependencies = ["pyyaml>=6.0"]

[tool.ruff]
target-version = "py310"
line-length = 120
